requirements.txt
Django==2.0.1
django-apptemplates==1.4
django-bootstrap3==9.1.0
django-braces==1.12.0
django-ckeditor==5.4.0
django-crispy-forms==1.7.0
django-debug-toolbar==1.9.1
django-js-asset==1.0.0
djangocms-admin-style==1.2.7
Faker==0.8.12
Pillow==5.0.0
psycopg2==2.7.3.2
python-dateutil==2.7.2
pytz==2017.3
six==1.11.0
sqlparse==0.2.4
text-unidecode==1.2
Unidecode==1.0.22
XlsxWriter==1.0.2
xlwt==1.3.0
